refactored sidebar components for organization. still getting bugs with joined clubs component returning undefined sometimes.
