+++ frontend/src/index.css
@@ -99,7 +99,6 @@ hr {
 #form__container {
     position: inherit;
     width: 50%;
-    height: 100%;
     display: flex;
     flex-direction: column;
     padding: 25px 60px;
